archive/old-services-20251020/main_migrated.py
#!/usr/bin/env python3
"""
Tower Anime_Production Service - Migrated to use Tower Common Library
"""

import sys
import os
from pathlib import Path

# Add tower-common to Python path
sys.path.insert(0, "/opt/tower-common")

from tower_common import TowerBaseService
from tower_common.errors import NotFoundError, ValidationError
from fastapi import APIRouter, Depends

class Anime_ProductionService(TowerBaseService):
    """Migrated Anime_Production service using Tower Common patterns"""

    def __init__(self):
        super().__init__(
            service_name="tower-anime-production",
            description="Anime production service with AI workflows",
            version="1.0.0",
            config_overrides={
                "port": 44451
            }
        )

        # Setup service-specific routes
        self._setup_routes()

    def _setup_routes(self):
        """Setup service-specific API routes"""
        router = APIRouter()

        # TODO: Migrate your existing routes here
        # Import your existing route handlers and add them to the router

        @router.get("/")
        async def root():
            """Root endpoint - replace with your existing logic"""
            return self.error_handlers.create_success_response(
                data={"message": "Welcome to Anime_Production Service"}
            )

        # Add the router to the service
        self.add_router(router, tags=["anime-production"])

    async def startup(self):
        """Service startup logic"""
        await super().startup()

        # TODO: Add your existing startup logic here
        # Examples from your service:
        # - Database initialization
        # - Model loading
        # - External service connections

        self.logger.get_logger().info(f"{self.service_name} migration startup completed")

    async def shutdown(self):
        """Service shutdown logic"""
        # TODO: Add your existing cleanup logic here

        await super().shutdown()

# Service factory
def create_service() -> Anime_ProductionService:
    """Create and configure the service"""
    return Anime_ProductionService()

if __name__ == "__main__":
    service = create_service()
    service.run()
