requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.5.0
requests>=2.31.0
aiohttp>=3.9.0
websockets>=12.0
python-multipart>=0.0.6
jinja2>=3.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
gtts>=2.4.0
pydub>=0.25.0